move utils outside of qrcode-modal and refactor web3-subprovider
