+++ packages/clients/client/tsconfig.json
@@ -1,5 +1,4 @@
 {
   "extends": "../../../tsconfig.json",
-
   "include": ["./src/**/*"]
 }
