+++ packages/helpers/browser-crypto/src/index.ts
@@ -12,7 +12,7 @@ import {
   convertHexToArrayBuffer,
   convertUtf8ToArrayBuffer,
   removeHexPrefix,
-  safeGetFromWindow,
+  getCrypto,
 } from "@walletconnect/utils";
 
 const AES_ALGORITHM = "AES-CBC";
@@ -20,7 +20,7 @@ const AES_LENGTH = 256;
 const HMAC_ALGORITHM = "SHA-256";
 
 export async function exportKey(cryptoKey: CryptoKey): Promise<ArrayBuffer> {
-  const browserCrypto = safeGetFromWindow<Crypto>("crypto");
+  const browserCrypto = getCrypto();
   const buffer: ArrayBuffer = await browserCrypto.subtle.exportKey("raw", cryptoKey);
   return buffer;
 }
@@ -37,14 +37,14 @@ export async function importKey(
   const algoParams: AesKeyAlgorithm | HmacImportParams =
     type === AES_ALGORITHM ? aesParams : hmacParams;
   const usages: string[] = type === AES_ALGORITHM ? ["encrypt", "decrypt"] : ["sign", "verify"];
-  const browserCrypto = safeGetFromWindow<Crypto>("crypto");
+  const browserCrypto = getCrypto();
   const cryptoKey = await browserCrypto.subtle.importKey("raw", buffer, algoParams, true, usages);
   return cryptoKey;
 }
 
 export async function generateKey(length?: number): Promise<ArrayBuffer> {
   const _length = length || 256;
-  const browserCrypto = safeGetFromWindow<Crypto>("crypto");
+  const browserCrypto = getCrypto();
   const cryptoKey = await browserCrypto.subtle.generateKey(
     {
       length: _length,
@@ -59,7 +59,7 @@ export async function generateKey(length?: number): Promise<ArrayBuffer> {
 
 export async function createHmac(data: ArrayBuffer, key: ArrayBuffer): Promise<ArrayBuffer> {
   const cryptoKey: CryptoKey = await importKey(key, "HMAC");
-  const browserCrypto = safeGetFromWindow<Crypto>("crypto");
+  const browserCrypto = getCrypto();
   const signature = await browserCrypto.subtle.sign(
     {
       length: 256,
@@ -94,7 +94,7 @@ export async function aesCbcEncrypt(
   iv: ArrayBuffer,
 ): Promise<ArrayBuffer> {
   const cryptoKey: CryptoKey = await importKey(key, AES_ALGORITHM);
-  const browserCrypto = safeGetFromWindow<Crypto>("crypto");
+  const browserCrypto = getCrypto();
   const result: ArrayBuffer = await browserCrypto.subtle.encrypt(
     {
       iv,
@@ -112,7 +112,7 @@ export async function aesCbcDecrypt(
   iv: ArrayBuffer,
 ): Promise<ArrayBuffer> {
   const cryptoKey: CryptoKey = await importKey(key, AES_ALGORITHM);
-  const browserCrypto = safeGetFromWindow<Crypto>("crypto");
+  const browserCrypto = getCrypto();
   const result: ArrayBuffer = await browserCrypto.subtle.decrypt(
     {
       iv,
