+++ packages/helpers/qrcode-modal/package.json
@@ -27,7 +27,7 @@
   "homepage": "https://github.com/WalletConnect/walletconnect-monorepo/tree/v1.0.0-beta",
   "license": "LGPL-3.0",
   "source": "src/index.ts",
-  "main": "dist/cjs/index.min.js",
+  "main": "dist/cjs/index.js",
   "types": "dist/cjs/index.d.ts",
   "unpkg": "dist/umd/index.min.js",
   "files": [
@@ -63,7 +63,8 @@
   },
   "dependencies": {
     "@walletconnect/mobile-registry": "^1.0.0-beta.95",
-    "detect-browser": "5.1.0",
+    "@walletconnect/types": "^1.0.0-beta.95",
+    "@walletconnect/utils": "^1.0.0-beta.95",
     "preact": "10.4.1",
     "qr-image": "3.2.0",
     "qrcode-terminal": "0.12.0"
