--- packages/helpers/qrcode-modal/src/browser/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from "./env";
-export * from "./url";
-export * from "./window";
