+++ packages/helpers/utils/package.json
@@ -58,6 +58,7 @@
   "dependencies": {
     "@ethersproject/address": "5.0.0-beta.134",
     "@walletconnect/types": "^1.0.0-beta.95",
+    "detect-browser": "5.1.0",
     "enc-utils": "2.1.0"
   },
   "gitHead": "165f7993c2acc907c653c02847fb02721052c6e7"
