+++ packages/providers/starkware-provider/tsconfig.json
@@ -1,5 +1,4 @@
 {
   "extends": "../../../tsconfig.json",
-
   "include": ["./src/**/*"]
 }
