+++ packages/providers/truffle-provider/package.json
@@ -52,9 +52,7 @@
     "webpack-cli": "3.3.10"
   },
   "dependencies": {
-    "@walletconnect/client": "^1.0.0-beta.95",
-    "@walletconnect/types": "^1.0.0-beta.95",
-    "qrcode-terminal": "0.12.0",
+    "@walletconnect/web3-subprovider": "^1.0.0-beta.95",
     "web3-provider-engine": "15.0.7"
   }
 }
