+++ packages/providers/web3-subprovider/package.json
@@ -54,5 +54,11 @@
     "webpack": "4.41.5",
     "webpack-cli": "3.3.10"
   },
+  "dependencies": {
+    "@walletconnect/client": "^1.0.0-beta.95",
+    "@walletconnect/qrcode-modal": "^1.0.0-beta.95",
+    "@walletconnect/types": "^1.0.0-beta.95",
+    "web3-provider-engine": "15.0.7"
+  },
   "gitHead": "165f7993c2acc907c653c02847fb02721052c6e7"
 }
